more test cases and more raise exception
i made one raise exception for the user if not giving the schema path in create command and solved the errors of it.
add one more test case for this condition.
there is still one problem needed to be solved that the attributes of the exception don't get updated with new parameters when calling it with new parameters